requirements.txt
attrs==21.2.0
Brotli==1.0.9
certifi==2021.5.30
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
dash==2.0.0
dash-bootstrap-components==0.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Fiona==1.8.20
Flask==2.0.1
Flask-Compress==1.10.1
geopandas==0.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
munch==2.5.0
numpy==1.21.2
pandas==1.3.3
plotly==5.3.1
pyproj==3.2.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
Shapely==1.7.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.0.1
